Don't attempt to copy missing file on reported success without output file
It seems that in some conditions, most likely with an empty tab or file,
Draw.io can exit with a successful (zero) exit status, but not actually
write a file. We should check to see if the cache file exists and only
attempt to copy it to the build directory if it exists.

Also moved away from logging an exception, as it's confusing. Instead
we make it clear that Draw.io may have misreported an exit status.

Fixes #8
Fixes #13