Introduce a Configuration type
We need to make an effort to type the BasePlugin.config dict, as we're
otherwise having to unpack and pass many arguments to some of the more
complicated callbacks.

I'm trying to straddle the line between a reasonable number of args and
overly tight coupling by only using this object in methods which
actually do processing for now, and not those involved in computing
configuration.