Export on-demand if we need {content}
There will be some performance hit doing this as we'll end up calling
ensure_file_cached() again for every diagram in
on_post_page(), and we'll call it for every reference here rather than
once per diagram, but it'll do for now. It may later make sense to
keep a set of already-referenced Sources so we don't repeat work.